+++ crates/zng/src/lib.rs
@@ -559,6 +559,7 @@ pub mod prelude {
     pub use crate::__prelude::*;
 }
 mod __prelude {
+    // TODO(breaking) add FrequencyUnits
     pub use crate::{color, gesture, keyboard, layout, mouse, task, timer, touch, widget};
 
     pub use zng_task::rayon::prelude::{